backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "fastapi==0.110.0",
    "feast[postgres]==0.49.0",
    "greenlet==3.2.2",
    "kafka-python>=2.2.11",
    "passlib[bcrypt]==1.7.4",
    "pydantic>=2.7.1",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.30",
    "uvicorn[standard]>=0.29.0",
]
